ignore race from runner
